Update Clowder extractor
