+++ .dockerignore
@@ -1,16 +1,3 @@
-# Ignore everything
-*
-
-# Allow files and directories
-!/api
-!*.py
-!/swagger
-!requirements.txt
-!manage
-
-# Ignore unnecessary files inside allowed directories
-# This should go after the allowed directories
-**/*~
-**/*.log
-**/.DS_Store
-**/Thumbs.db
+data/resources
+data/WBD_National_GDB
+volumes
