+++ .editorconfig
@@ -18,7 +18,3 @@ max_line_length = 119
 # The JSON files contain newlines inconsistently
 [*.json]
 insert_final_newline = false
-
-# Makefiles always use tabs for indentation
-[Makefile]
-indent_style = tab
