+++ .env-example
@@ -1,39 +1,28 @@
-# ----------------------------------------------------------------------
-# FLASK/IMPORT_DATA.PY CONFIG
-# ----------------------------------------------------------------------
+TRAEFIK_HOST=localhost
+TRAEFIK_HTTP_PORT=80
+TRAEFIK_HTTPS_PORT=443
 
-DATABASE_URI="postgresql+psycopg2://user:pass@localhost:5432/gltg_bmp"
-DEBUG=true
+API_DEBUG=true
 API_PORT=8000
-LOG_LEVEL=INFO
+API_LOG_LEVEL=INFO
+API_CONTEXT=/bmp-api
 
-# ----------------------------------------------------------------------
-# ROUTES CONFIGURATION
-# ----------------------------------------------------------------------
-
-API_CONTEXT=/api/v1.0
+CLOWDER_URL=http://clowder:9000
 CLOWDER_CONTEXT=/clowder
-RABBITMQ_CLOWDERURL=http://clowder:9000/clowder
-
-# ----------------------------------------------------------------------
-# DATABASE CONFIGURATION
-# ----------------------------------------------------------------------
-
-DB_USER=postgres
-DB_PASSWORD=supersecretpassword
-DB_NAME=gltg_bmp
-DB_PORT=5432
-
-# ----------------------------------------------------------------------
-# TRAEFIK CONFIGURATION
-# ----------------------------------------------------------------------
-
-#TRAEFIK_ACME_EMAIL=abc@example.com
-#TRAEFIK_ACME_ENABLE=true
-
-# ----------------------------------------------------------------------
-# CLOWDER CONFIGURATION
-# ----------------------------------------------------------------------
-
+CLOWDER_MONITOR=/monitor
+CLOWDER_SSL=false
 CLOWDER_REGISTER=true
 CLOWDER_ADMINS=admin@example.com
+
+RABBITMQ_EXCHANGE=clowder
+RABBITMQ_DEFAULT_USER=guest
+RABBITMQ_DEFAULT_PASS=guest
+
+SMTP_MOCK=true
+SMTP_SERVER=smtp
+
+DB_HOST=postgres
+DB_PORT=5432
+DB_USER=postgres
+DB_PASSWORD=123456
+DB_NAME=gltg_bmp
