+++ .pre-commit-config.yaml
@@ -1,22 +1,22 @@
 repos:
-    - repo: https://github.com/asottile/reorder_python_imports
-      rev: v2.3.5
-      hooks:
-        - id: reorder-python-imports
-          name: Reorder Python imports (src, tests)
-    - repo: https://github.com/python/black
-      rev: stable
-      hooks:
-        - id: black
-    - repo: https://gitlab.com/pycqa/flake8
-      rev: 3.8.3
-      hooks:
-        - id: flake8
-          args: [--max-line-length=120]
-          additional_dependencies: [flake8-bugbear]
-    - repo: https://github.com/pre-commit/pre-commit-hooks
-      rev: v3.2.0
-      hooks:
-        - id: check-byte-order-marker
-        - id: trailing-whitespace
-        - id: end-of-file-fixer
+  - repo: https://github.com/pycqa/isort
+    rev: 5.8.0
+    hooks:
+      - id: isort
+        args: [ "--profile", "black", "--filter-files" ]
+  - repo: https://github.com/python/black
+    rev: 21.5b2
+    hooks:
+      - id: black
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: 3.9.2
+    hooks:
+      - id: flake8
+        args: [--max-line-length=120]
+        additional_dependencies: [flake8-bugbear]
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.0.1
+    hooks:
+      - id: check-byte-order-marker
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
