--- api/db.py
@@ -1,18 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import scoped_session
-from sqlalchemy.orm import sessionmaker
-
-
-Base = declarative_base()
-
-
-def get_db(db_uri):
-    engine = create_engine(db_uri, convert_unicode=True)
-    db_session = scoped_session(
-        sessionmaker(autocommit=False, autoflush=False, bind=engine)
-    )
-
-    Base.query = db_session.query_property()
-
-    return db_session
