+++ api/swagger/v1.0.yaml
@@ -1,8 +1,5 @@
 openapi: 3.0.3
 
-servers:
-  - url: /api/v1.0
-
 info:
   title: Best Management Practices Portal API
   version: '1.0'
