+++ api/utils/cli.py
@@ -1,0 +1,2 @@
+def is_true(value: str) -> bool:
+    return value.lower() in ("true", "1", "t")
