--- bmp_practices_extractor/README.md
@@ -1,3 +1,0 @@
-## BMP Practices Extractor
-
-This extractor is based on the [Simple Extractor](https://github.com/clowder-framework/pyclowder/blob/b852c63523b769baee74c00cf12fa9c7d77f963d/pyclowder/extractors.py#L293) provided in pyclowder. It uses the import_data script to get and process BMP practices data from excel files uploaded to a clowder instance.
