--- bmp_practices_extractor/requirements.txt
@@ -1,8 +1,0 @@
-pyclowder==2.3.4
-SQLAlchemy==1.3.19
-GeoAlchemy2==0.8.4
-python-dotenv==0.14.0
-geopandas==0.8.1
-pandas==1.1.2
-psycopg2-binary==2.8.6
-xlrd
