+++ docker/bmp_api/Dockerfile
@@ -1,0 +1,9 @@
+FROM python:3.9-slim
+
+COPY ./requirements /tmp/requirements
+RUN pip install -r /tmp/requirements/api.txt
+RUN rm -r /tmp/requirements
+
+WORKDIR /home/gltg-bmp
+
+ENTRYPOINT ["python", "api/app.py"]
