--- manage
@@ -1,32 +1,0 @@
-#!/usr/bin/env sh
-
-help() {
-    echo "Usage"
-    echo "-h             show this help"
-    echo "-d    --dev    set the environment to development and debug to true for the flask app"
-    exit
-}
-
-export FLASK_APP=app.py
-export FLASK_DEBUG=0
-export FLASK_ENV=production
-
-if [ "$DEBUG" ]; then
-  FLASK_DEBUG=1
-  FLASK_ENV=development
-fi
-
-for i in "$@"
-do
-    case $i in
-    -h)
-    help
-    ;;
-    -d|--dev)
-    FLASK_DEBUG=1
-    FLASK_ENV=development
-    ;;
-    esac
-done
-
-python app.py
