--- requirements.txt
@@ -1,10 +1,0 @@
-Flask==1.1.2
-Flask-Cors==3.0.9
-SQLAlchemy==1.3.20
-GeoAlchemy2==0.8.4
-connexion[swagger-ui]==2.7.0
-python-dotenv==0.15.0
-geopandas==0.8.1
-pandas==1.1.4
-psycopg2-binary==2.8.6
-xlrd==1.2.0
