+++ requirements/common.txt
@@ -1,0 +1,7 @@
+SQLAlchemy==1.4.18
+GeoAlchemy2==0.9.0
+geopandas==0.9.0
+openpyxl==3.0.7
+pandas==1.2.4
+psycopg2-binary==2.8.6
+python-dotenv==0.17.1
