+++ requirements/dev.txt
@@ -1,4 +1,3 @@
 black
 ipython
 pre-commit
-xlrd
