+++ requirements/extractor.txt
@@ -1,0 +1,2 @@
+-r common.txt
+pyclowder==2.4.0
